0.1.2
Changed default window size so as to fit into a 1024 x 768 screen size.
Change default compatibility to Scintilla POSIX, because it's faster in general.
Change logic to always display the Replace field, but make it disabled when compatibility is set to Scintilla.
Updated readme.md to include download release location.
Added usage instructions in the default sample text.